requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.0
Flask==3.0.3
Flask-Session==0.8.0
Flask-WTF==1.2.1
idna==3.8
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
lxml==5.3.0
lxml_html_clean==0.2.2
MarkupSafe==2.1.5
msgspec==0.18.6
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.1.1
pandas==2.2.2
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
scikit-learn==1.5.1
scipy==1.14.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.6
threadpoolctl==3.5.0
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.5
tzdata==2024.1
urllib3==2.2.2
validators==0.34.0
Werkzeug==3.0.4
WTForms==3.1.2
